Change Enum Location
