+++ src/index.ts
@@ -3,7 +3,7 @@ import getMonthlyLeaderboard from "./methods/getMonthlyLeaderboard";
 import getAllTimeStats from "./methods/getAllTimeStats";
 import getAllTimeLeaderboard from "./methods/getAllTimeLeaderboard";
 import getGlobalStatistics from "./methods/getGlobalStatistics";
-import { GAME as GAMES } from "./types/GAMES";
+import { GAME as GAMES } from "./types/GAME_INFO";
 
 export {
     getAllTimeStats,
